requirements.txt
absl-py==0.7.1
aiml==0.9.2
altgraph==0.16.1
asn1crypto==0.24.0
astor==0.7.1
astroid==2.0.4
attrs==19.1.0
Automat==0.7.0
autopep8==1.4.4
blis==0.2.4
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
ChatterBot==1.0.5
chatterbot-corpus==1.2.0
Click==7.0
clipspy==0.3.2
colorama==0.4.0
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
cycler==0.10.0
cymem==2.0.2
Flask==1.0.2
Flask-SocketIO==3.3.2
future==0.17.1
gast==0.2.2
get==2019.4.13
grpcio==1.20.1
gunicorn==19.9.0
h5py==2.9.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
line-bot-sdk==1.8.0
lxml==4.3.4
Markdown==3.1
MarkupSafe==1.1.1
mathparse==0.1.2
matplotlib==3.1.1
mccabe==0.6.1
mock==3.0.3
murmurhash==1.0.2
mysql-connector-python==8.0.16
nltk==3.4
numpy==1.16.2
oauthlib==3.0.1
opencv-python==4.1.1.26
parsel==1.5.1
pefile==2019.4.18
Pillow==6.0.0
Pint==0.9
pipenv==2018.11.26
plac==0.9.6
post==2019.4.13
preshed==2.0.1
protobuf==3.7.1
public==2019.4.13
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyInstaller==3.5
pylint==2.1.1
pymongo==3.7.2
pyOpenSSL==19.0.0
pyparsing==2.4.5
pypiwin32==223
PySastrawi==1.2.0
PySocks==1.6.8
python-dateutil==2.7.5
python-decouple==3.1
python-engineio==3.5.1
python-socketio==3.1.2
pytz==2019.1
pywin32==224
pywin32-ctypes==0.2.0
PyYAML==3.13
query-string==2019.4.13
queuelib==1.5.0
request==2019.4.13
requests==2.21.0
requests-oauthlib==1.2.0
scikit-learn==0.20.3
scipy==1.2.1
Scrapy==1.6.0
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
spacy==2.1.3
SQLAlchemy==1.2.18
srsly==0.0.5
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
thinc==7.0.4
tqdm==4.31.1
tweepy==3.7.0
Twisted==19.2.1
urllib3==1.24.1
virtualenv==16.4.3
virtualenv-clone==0.5.3
w3lib==1.20.0
wasabi==0.2.1
Werkzeug==0.15.2
wrapt==1.10.11
zope.interface==4.6.0
